Rename the file name
