+++ src/test/java/JenkinsTests/IntegrationTest.java
@@ -1,4 +1,4 @@
-package MyTest;
+package JenkinsTests;
 
 import com.cloudbees.plugins.credentials.CredentialsMatchers;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
