+++ src/test/java/JenkinsTests/LoadConfigureTest.java
@@ -1,0 +1,22 @@
+package JenkinsTests;
+
+import hudson.XmlFile;
+import hudson.model.FreeStyleProject;
+import org.junit.Test;
+
+import java.util.logging.Logger;
+
+/**
+ * Created by t-yuhang on 7/31/2017.
+ */
+public class LoadConfigureTest extends IntegrationTest {
+
+    private static final Logger LOGGER = Logger.getLogger(LoadConfigureTest.class.getName());
+
+    @Test
+    public void LoadTest() throws Exception {
+
+    }
+
+
+}
